[ESPP] Update regarding ILS -> USD exchange rate
Per finance in our company, the conversion is done at the end of
the period, and is based on the 6-month trailing average exchange
rate between USD to ILS (which is similar to what is done for RSU
grants when entering the company). This is logically equivalent to
a gradual conversion, yet the update is done so to maintain
correctness.

Signed-off-by: Eyal Itkin <eyal.itkin@hotmail.com>